Revert "Split Serial module into Serial and WSerial"
This reverts commit 5ddfb4576482b2ec712fffa0771b114c3e821958.
