+++ streamly.cabal
@@ -344,7 +344,6 @@ library
                      , Streamly.Internal.Data.Stream.SVar.Eliminate
 
                      , Streamly.Internal.Data.Stream.Serial
-                     , Streamly.Internal.Data.Stream.WSerial
                      , Streamly.Internal.Data.Stream.Async
                      , Streamly.Internal.Data.Stream.Parallel
                      , Streamly.Internal.Data.Stream.Channel.Types
